split in to utils and configs for webpack cyclic explosion
