+++ dev-utils/test-config.js
@@ -23,14 +23,55 @@
  *
  */
 
-module.exports = function() {
-  return (
-    window.globalConfigs &&
-    (window.globalConfigs.agentConfig || {
-      serverUrl: 'http://localhost:8200',
-      serviceName: 'test',
-      agentName: 'apm-js-core',
+function getSauceConnectOptions() {
+  return {
+    username: process.env.SAUCE_USERNAME,
+    accessKey: process.env.SAUCE_ACCESS_KEY,
+    logger: console.log,
+    noSslBumpDomains: 'all',
+    tunnelIdentifier: process.env.TRAVIS_JOB_NUMBER,
+    connectRetries: 3
+  }
+}
+
+function getTestEnvironmentVariables() {
+  return {
+    branch: process.env.TRAVIS_BRANCH,
+    mode: process.env.MODE,
+    sauceLabs: process.env.MODE && process.env.MODE.startsWith('saucelabs'),
+    isTravis: process.env.TRAVIS,
+    serverUrl: process.env.APM_SERVER_URL || 'http://localhost:8200'
+  }
+}
+
+function getGlobalConfig(packageName = 'rum') {
+  const env = getTestEnvironmentVariables()
+  const globalConfigs = {
+    agentConfig: {
+      serverUrl: env.serverUrl,
+      serviceName: `test`,
+      agentName: `${packageName}`,
       agentVersion: '0.0.1'
-    })
-  )
+    },
+    useMocks: false,
+    mockApmServer: false
+  }
+
+  /**
+   * Use this for testing locally
+   */
+  // if (env.sauceLabs) {
+  //   globalConfigs.useMocks = true
+  // }
+
+  return {
+    globalConfigs,
+    testConfig: env
+  }
+}
+
+module.exports = {
+  getSauceConnectOptions,
+  getTestEnvironmentVariables,
+  getGlobalConfig
 }
