+++ packages/rum/test/e2e/general-usecase/webpack.config.js
@@ -26,9 +26,9 @@
 const path = require('path')
 const webpack = require('webpack')
 const UglifyJSPlugin = require('uglifyjs-webpack-plugin')
-const { getTestConfig } = require('../../../../../dev-utils/test')
+const { getGlobalConfig } = require('../../../../../dev-utils/test-config')
 
-const { globalConfigs } = getTestConfig()
+const { globalConfigs } = getGlobalConfig()
 
 const configJson = JSON.stringify(globalConfigs, undefined, 2)
 const env = { globalConfigs: configJson }
